[hack] Squish the KL
The original one gave a cross-entropy of about 1e6 - if we take the mean, that
gives ~1. This is just a way of manually specifying the weight we put on the
mask reconstruction. So even if we don't get this exactly the same as the
principled version, we're adjusting the 'gamma' weight to get it to train.

Help me implement this better